[improvement][debate-structures] Move 8 scripted conversation patterns from swarms/structs into examples
InterviewSeries, PeerReviewProcess, MediationSession, BrainstormingSession,
TrialSimulation, CouncilMeeting, MentorshipSession, and NegotiationSession
move from swarms/structs/multi_agent_debates.py to
examples/multi_agent/alternate_debates/, one self-contained file each.

They are compositions, not framework machinery: each is built entirely from
the public Agent, Conversation, and history_output_formatter APIs, adds no
capability to the framework, and is most useful as something you copy and
adapt. None was exported from swarms/__init__.py, so there is no public API
change.

Each new file carries the class plus a runnable __main__ demo, so
`python examples/multi_agent/alternate_debates/council_meeting.py` works
directly. A README documents all eight, the six-step template most of them
share, and the known TrialSimulation cross-examination defect.

Also:
- Relocated the 34 tests for these classes to
  examples/multi_agent/alternate_debates/test_alternate_debates.py rather
  than deleting them, so coverage travels with the code.
- Repointed the 8 examples in orchestration_examples/ that imported these
  classes at their new location.

swarms/structs/multi_agent_debates.py: 1192 -> 280 lines.
