+++ examples/multi_agent/orchestration_examples/cybersecurity_incident_negotiation.py
@@ -1,5 +1,12 @@
 from swarms import Agent
-from swarms.structs.multi_agent_debates import NegotiationSession
+import sys
+from pathlib import Path
+
+sys.path.insert(
+    0, str(Path(__file__).resolve().parents[1] / "alternate_debates")
+)
+
+from negotiation_session import NegotiationSession  # noqa: E402
 
 # Initialize the negotiation participants
 incident_mediator = Agent(
