+++ examples/multi_agent/orchestration_examples/investment_council_meeting.py
@@ -1,5 +1,12 @@
 from swarms import Agent
-from swarms.structs.multi_agent_debates import CouncilMeeting
+import sys
+from pathlib import Path
+
+sys.path.insert(
+    0, str(Path(__file__).resolve().parents[1] / "alternate_debates")
+)
+
+from council_meeting import CouncilMeeting  # noqa: E402
 
 # Initialize the investment council members
 investment_chair = Agent(
