+++ examples/multi_agent/orchestration_examples/medical_malpractice_trial.py
@@ -1,5 +1,12 @@
 from swarms import Agent
-from swarms.structs.multi_agent_debates import TrialSimulation
+import sys
+from pathlib import Path
+
+sys.path.insert(
+    0, str(Path(__file__).resolve().parents[1] / "alternate_debates")
+)
+
+from trial_simulation import TrialSimulation  # noqa: E402
 
 # Initialize the trial participants
 prosecution_attorney = Agent(
