+++ examples/multi_agent/orchestration_examples/startup_mentorship_program.py
@@ -1,5 +1,12 @@
 from swarms import Agent
-from swarms.structs.multi_agent_debates import MentorshipSession
+import sys
+from pathlib import Path
+
+sys.path.insert(
+    0, str(Path(__file__).resolve().parents[1] / "alternate_debates")
+)
+
+from mentorship_session import MentorshipSession  # noqa: E402
 
 # Initialize the mentor and mentee
 startup_mentor = Agent(
